+++ riverpie_flutter/lib/addons.dart
@@ -1,0 +1,2 @@
+export 'package:riverpie_flutter/src/addons/action.dart';
+export 'package:riverpie_flutter/src/addons/snack_bar.dart';